Rename `our_stdint.h` to `stdint_compat.h`
